Change the order on which keys are deleted and stored for safety.
This change switches the order of the keychain migration so that
keys are first stored in the new value securely and then deleted
from the legacy keys. This requires also handling the "extra wipe"
in a different function so that the new keys are not deleted as
part of the "cleanup".